Changed fabfile name.
