--- deploy/fab/ec2-deploy.sh
@@ -1,1 +1,0 @@
-USE_LATEST_KEY='True' EC2_INSTANCE_TYPE='m3.large' SERVER_FILE_PATH="XXX" PUF_FILE_PATH="XXX" GITHUB_USERNAME='ospc-auto' GITHUB_PRIVATE_KEY_PATH='XXX/ospc-auto/id_rsa' OGUSA_GIT_BRANCH='XXX' DROPQ_GIT_BRANCH='XXX' DEPLOY_GIT_BRANCH='XXX' AWS_ID='XXX' AWS_SECRET='XXX' python ec2-fabfile.py
