+++ deploy/fab/ec2_fabfile.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 from __future__ import print_function
+from functools import partial
 import os
 import sys
 import boto.exception
@@ -29,10 +30,16 @@ from fabric.contrib.files import (
 )
 
 import logging
+from copy_deploy_repo import copy_deploy_repo, check_unmodified
+from redeploy import cli as deploy_versions_cli, main as reset_server_main
 
-from copy_deploy_repo import copy_deploy_repo
+DEPLOYMENT_VERSIONS_ARGS = deploy_versions_cli(ip_address='skip')
+if not 'OSPC_ANACONDA_TOKEN' in os.environ:
+    raise ValueError('OSPC_ANACONDA_TOKEN must be defined in env vars')
+check_unmodified()
 #based on https://github.com/ContinuumIO/wakari-deploy/blob/master/ami_creation/fabfile.py
-
+ssh_transport = logging.getLogger("ssh.transport")
+ssh_transport.disabled = True
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
 formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
@@ -45,15 +52,6 @@ console.setFormatter(formatter)
 # log.addHandler(handler)
 log.addHandler(console)
 
-# ami-65ce000e
-# ami-849e0c84
-# ami-1998e77c
-# ami-af5591eb
-# ami-9f92edfa
-# ami-1d4b1f78
-#AMI_ID = 'ami-1d4b1f78' #SSD in us-east-1
-#AMI_ID = 'ami-dfcab0b5' #SSD in us-east-1
-#AMI_ID = 'ami-dfcab0b5' #SSD in us-east-1
 AMI_ID = 'ami-6c276206' #SSD in us-east-1
 
 use_latest_key = 'True' == os.environ.get('USE_LATEST_KEY', False)
@@ -72,30 +70,27 @@ if 'xxx' == os.environ.get('SERVER_FILE_PATH', 'xxx').lower():
     quit()
 
 else:
-    PUF_FILE_PATH = os.path.abspath(os.path.expanduser(os.environ.get('PUF_FILE_PATH')))
     SERVER_FILE_PATH = os.path.abspath(os.path.expanduser(os.environ.get('SERVER_FILE_PATH')))
-    if not os.path.exists(PUF_FILE_PATH):
-        print("{} does not exist".format(PUF_FILE_PATH))
-        quit()
 
+ec2 = boto.ec2.connect_to_region('us-east-1', #N. Virginia
+        aws_access_key_id = os.environ.get('AWS_ID'),
+        aws_secret_access_key = os.environ.get('AWS_SECRET'))
 env.use_ssh_config = True
 env.disable_known_hosts = True
 env.connection_attempts = True
-env.timeout = 300
+env.timeout = 50
 
 KEYNAME = "{}-aei-dropq-{}".format(AMI_ID,os.environ.get('USER', 'ospc'))
-#KEYNAME = "keypair-ospc"
-#SECURITY_GROUP = 'launch-wizard-26'
-#SECURITY_GROUP = 'dropq-security-group'
 SECURITY_GROUP = 'dropq-iam-group'
+NODE_COUNT = int(os.environ.get('WORKER_NODE_COUNT', 1))
 
 def create_box():
     old_ids = set(i.id for i in ec2.get_only_instances())
-    machine = ec2.run_instances(AMI_ID, key_name=KEYNAME, min_count=2, max_count=2,
-        security_groups=[SECURITY_GROUP,], instance_type=os.environ.get('EC2_INSTANCE_TYPE', 'm3.medium'))
+    machine = ec2.run_instances(AMI_ID, key_name=KEYNAME, min_count=NODE_COUNT, max_count=NODE_COUNT,
+        security_groups=[SECURITY_GROUP,], instance_type=os.environ.get('EC2_INSTANCE_TYPE', 'r3.large'))
     new_instances = [i for i in ec2.get_only_instances() if i.id not in old_ids]
     for new_instance in new_instances:
-        print(new_instance.id)
+        print("new instance:", new_instance.id)
         ec2.create_tags([new_instance.id], {"billingProject": "aei"})
 
 
@@ -118,7 +113,7 @@ def create_box():
     time.sleep(1)
     for new_instance in new_instances:
         assert new_instance.public_dns_name
-        print(new_instance.public_dns_name)
+        print("public dns name:", new_instance.public_dns_name)
 
     return new_instances
 
@@ -154,6 +149,7 @@ def test_all_ssh(instances, key_file):
     #env.hosts = ['ubuntu@e' + ip for ip in ips]
     env.hosts = [ip for ip in ips]
     env.user = 'ubuntu'
+    env.password = ''
     env.key_file = key_file
     env.key_filename = key_file
 
@@ -164,18 +160,36 @@ def test_all_ssh(instances, key_file):
 
     log.info('Key file: %s' % (key_file))
     log.debug('Trying to connect...')
+    import pdb;pdb.set_trace()
     for h in env.hosts:
         env.host_string = h
         run('pwd')
 
 def test_ssh2():
+    '''log.info('Key file: %s' % (key_file))
+    log.debug('Trying to connect...')
+    for ip in ips:
+        env.host = 'ubuntu@' + ip
+        env.password = ''
+        '''
     run('pwd')
+
+def add_public_keys():
+    # Add the necessary public key information to the instance.
+    from boto.s3.connection import S3Connection
+    conn = S3Connection(os.environ.get('AWS_ID'), os.environ.get('AWS_SECRET'))
+    bucket = conn.get_bucket("taxbrain-deploy-assets")
+
+    k = bucket.get_key("keys.txt")
+    keyfile = k.get_contents_as_string()
+    run("echo $'%s' >> ~/.ssh/authorized_keys" % keyfile)
 
 def test_ssh(instance, key_file):
     # needed to convert from unicode to ascii?
     key_file = str(key_file)
     ip = str(instance.public_dns_name)
     env.host = 'ubuntu@e' + ip
+    env.password = ''
     env.host_string = ip
     env.hosts = [env.host]
     env.user = 'ubuntu'
@@ -191,6 +205,7 @@ def test_ssh(instance, key_file):
 
 def connect_to_existing_machine(ip, key_file_path):
     env.user = 'ubuntu'
+    env.password = ''
     env.hosts = ['{}@{}'.format(env.user, ip)]
     env.host = '{}@{}'.format(env.user, ip)
     env.host_string = '{}@{}'.format(env.user, ip)
@@ -216,29 +231,25 @@ def apt_installs():
     sudo("apt-get install -y {}".format(' '.join(packages)))
 
 
-def install_ogusa_repo():
-    if 'XXX' not in os.environ.get('GITHUB_PRIVATE_KEY_PATH'):
-        with settings(warn_only = True):
-            run('ssh -T git@github.com -o StrictHostKeyChecking=no')
-        url = 'git@github.com:open-source-economics/OG-USA'
-    else:
-      url = 'https://{USERNAME}@github.com/open-source-economics/OG-USA'.format(
-          USERNAME = os.environ.get('GITHUB_USERNAME'))
+def install_deploy_repo():
+    sudo('rm -rf ~/deploy')
+    run('mkdir ~/deploy')
+    copy_deploy_repo(sudo, put, run)
 
+
+def install_ogusa_repo():
+    url = 'https://github.com/open-source-economics/OG-USA'
     sudo('rm -rf ~/OG-USA')
-    if os.environ.get('GIT_BRANCH'):
+    if os.environ.get('OGUSA_GIT_BRANCH'):
         run("git clone {} --branch {}".format(url, os.environ.get('OGUSA_GIT_BRANCH')))
     else:
         run("git clone {}".format(url))
 
 
+def write_ospc_anaconda_token():
+    token = os.environ['OSPC_ANACONDA_TOKEN']
+    run('echo {token} &> /home/ubuntu/.ospc_anaconda_token'.format(token=token))
 
-def copy_puf_file():
-    put(PUF_FILE_PATH, "/home/ubuntu/deploy/puf.csv.gz")
-    put(SERVER_FILE_PATH, "/home/ubuntu/reset_server.sh")
-
-def extract_puf_file():
-    run("cd /home/ubuntu/deploy/; gunzip -k puf.csv.gz")
 
 def convenience_aliases():
     run('echo "alias supervisorctl=\'supervisorctl -c /home/ubuntu/deploy/fab/supervisord.conf\'" >> ~/.bashrc')
@@ -246,37 +257,111 @@ def convenience_aliases():
 
 
 def run_salt():
-    run("ln -s ~/deploy/salt ~/salt")
-    sudo('sudo /usr/bin/salt-call state.highstate --retcode-passthrough --log-level=debug --config-dir="$HOME/deploy/salt"')
+    run("ln -s ~/deploy/fab/salt ~/salt")
+    sudo('sudo /usr/bin/salt-call state.highstate --retcode-passthrough --log-level=debug --config-dir="$HOME/deploy/fab/salt"')
+
+
+def _env_str(args):
+    s = []
+    for k in dir(args):
+        val = getattr(args, k)
+        if any(tok in k.lower() for tok in ('install', 'method' 'version')):
+            s.append('{}="{}"'.format(k.upper(), val))
+    return " ".join(s)
 
 def reset_server():
-    run("source /home/ubuntu/reset_server.sh")
+    e = _env_str(DEPLOYMENT_VERSIONS_ARGS)
+    command = "{} source /home/ubuntu/reset_server.sh".format(e)
+    print('Running', command)
+    run(command, shell=True, shell_escape=False)
 
 
-key_filename = './latest.pem'
-#instances = ['ip1', 'ip2']
+def wait_for_login_prompt(instance, retry_count=20):
+    import time
+    import re
+
+    for i in range(retry_count):
+        print("checking for login prompt on", instance, "at", time.ctime())
+        console = instance.get_console_output()
+        try:
+            mo = re.search("login:", console.output)
+            if mo is not None:
+                #We got the login console for this machine, so we are done.
+                print("Got login console for", instance)
+                break
+
+        except TypeError:
+            #We'll get a TypeError if the console output is None
+            pass
+        time.sleep(30)
+
+    if i == retry_count:
+        #If we retried our maximum number of times, indicate failure
+        raise RuntimeError("Instance was not up after trying for {} retries".format(retry_count))
+
+if os.environ.get('DROPQ_IP_ADDR'):
+    ip_address = os.environ.get('DROPQ_IP_ADDR')
+    key_filename = os.path.abspath('./keys/ec2-{}.pem'.format(ip_address))
+    if not os.path.exists(key_filename):
+        key_filename = os.path.abspath('latest.pem')
+    public_dns_name = 'ec2-{}.compute-1.amazonaws.com'.format(ip_address.replace('.','-'))
+    connect_to_existing_machine(ip_address, key_filename)
+else:
+    instances = create_box()
+    public_dns_names = []
+    ip_addresses = []
+    keyfile = "./" + KEYNAME + ".pem"
+    for instance in instances:
+        print("trying this instance :", instance)
+        subprocess.check_output(['cp', keyfile,
+                                os.path.join(os.path.dirname(keyfile), 'ec2-{}.pem'.format(instance.ip_address))])
+        public_dns_names.append(instance.public_dns_name)
+        ip_addresses.append(instance.ip_address)
+
+if 'quitafterec2spinup' in sys.argv:
+    quit()
+
+
 key_file = str(key_filename)
-ips = instances
-env.hosts = [ip for ip in ips]
+fqdns = [str(instance.public_dns_name) for instance in instances]
+env.hosts = [fqdn for fqdn in fqdns]
 env.user = 'ubuntu'
+env.password = ''
 env.key_file = key_file
 env.key_filename = key_file
 # forward ssh agent -- equivalent of ssh -A
 env.forward_agent = True
 
-print(env.hosts)
-print(ips)
-#execute(test_ssh2)
-#execute(apt_installs)
-#execute(install_ogusa_repo)
-#execute(convenience_aliases)
-#execute(copy_puf_file)
-#execute(extract_puf_file)
+print("hosts:", env.hosts)
+print("ips:", fqdns)
+
+#Wait for the machines to actually come up:
+
+for instance in instances[:]:
+    try:
+        wait_for_login_prompt(instance)
+    except RuntimeError:
+        #This host isn't ready, so terminate it and cleanup
+        env.hosts.remove(str(instance.public_dns_name))
+        instances.remove(instance)
+        print("Terminating", instance)
+        instance.terminate()
+
+
+execute(test_ssh2)
+execute(add_public_keys)
+execute(apt_installs)
+execute(install_deploy_repo)
+execute(install_ogusa_repo)
+execute(convenience_aliases)
+execute(write_ospc_anaconda_token)
 execute(run_salt)
 execute(reset_server)
 
 for instance in instances:
-    ssh_command = 'ssh -i {key} ubuntu@{ip} "'.format(ip=instance, key=key_filename)
-    with open("log_{}.log".format(instance), 'w') as f:
+    ssh_command = 'ssh -i {key} ubuntu@{ip} "'.format(ip=instance.ip_address,
+            key=KEYNAME)
+    with open("log_{}.log".format(instance.ip_address), 'w') as f:
         f.write(ssh_command)
     print(ssh_command)
+
